LSP design principle changes.
