+++ src/design/principles/LSP/LSPFollowed.java
@@ -84,6 +84,20 @@ class FDAccount implements DepositOnlyAccount {
     }
 }
 
+record Client(List<DepositOnlyAccount> fdAccounts,
+              List<WithDrawableAccount> drawableAccounts) {
+    public void processFDTransactions() {
+        fdAccounts.forEach(account -> account.deposit(500));
+    }
+
+    public void processWithDrawableTransactions() {
+        drawableAccounts.forEach(account -> {
+            account.deposit(500);
+            account.withdraw(200);
+        });
+    }
+}
+
 public class LSPFollowed {
     public static void main(String[] args) {
