+++ src/design/principles/LSP/LSPViolated.java
@@ -81,6 +81,21 @@ class FixedDepositAccount implements Account {
     }
 }
 
+record BankClient(List<Account> accounts) {
+
+    public void performTransactions() {
+        accounts.forEach(account -> {
+            System.out.println("\nUsing account: " + account.getClass().getSimpleName());
+            account.deposit(500);
+            try {
+                account.withdraw(200);
+            } catch (UnsupportedOperationException e) {
+                System.out.println("Exception : " + e.getMessage());
+            }
+        });
+    }
+}
+
 public class LSPViolated {
     public static void main(String[] args) {
