+++ src/design/principles/LSP/LSPWronglyHandled.java
@@ -2,6 +2,21 @@ package design.principles.LSP;
 
 import java.util.List;
 
+record BanksClient(List<Account> accounts) {
+    public void performTransactions() {
+        accounts.forEach(account -> {
+            System.out.println("\nUsing account: " + account.getClass().getSimpleName());
+            if (account instanceof FixedDepositAccount) {
+                account.deposit(500);
+                System.out.println("Skipping withdraw operation for FixedDepositAccount.");
+            } else {
+                account.deposit(500);
+                account.withdraw(200);
+            }
+        });
+    }
+}
+
 public class LSPWronglyHandled {
     public static void main(String[] args) {
